@@ -13,44 +13,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/schematics": "^10.2.3",
-    "@angular/animations": "10.0.14",
-    "@angular/cdk": "^10.1.3",
-    "@angular/common": "10.0.14",
-    "@angular/compiler": "10.0.14",
-    "@angular/core": "10.0.14",
-    "@angular/forms": "10.0.14",
-    "@angular/http": "7.2.16",
-    "@angular/localize": "10.0.14",
-    "@angular/platform-browser": "10.0.14",
-    "@angular/platform-browser-dynamic": "10.0.14",
-    "@angular/platform-server": "10.0.14",
-    "@angular/router": "10.0.14",
+    "@angular-devkit/schematics": "^13.2.0",
+    "@angular/animations": "13.2.0",
+    "@angular/common": "13.2.0",
+    "@angular/compiler": "13.2.0",
+    "@angular/core": "13.2.0",
+    "@angular/forms": "13.2.0",
+    "@angular/localize": "13.2.0",
+    "@angular/platform-browser": "13.2.0",
+    "@angular/platform-browser-dynamic": "13.2.0",
+    "@angular/platform-server": "13.2.0",
+    "@angular/router": "13.2.0",
     "@babel/plugin-proposal-class-properties": "^7.13.0",
     "@microsoft/signalr": "^5.0.6",
-    "@pdftron/webviewer-video": "^3.0.13",
-    "@progress/kendo-angular-buttons": "^5.5.2",
-    "@progress/kendo-angular-common": "^1.2.4",
-    "@progress/kendo-angular-dateinputs": "^4.4.1",
-    "@progress/kendo-angular-dialog": "^4.2.3",
-    "@progress/kendo-angular-dropdowns": "^4.4.1",
-    "@progress/kendo-angular-editor": "^1.2.5",
-    "@progress/kendo-angular-excel-export": "^3.1.6",
-    "@progress/kendo-angular-grid": "^4.8.4",
-    "@progress/kendo-angular-inputs": "^6.7.4",
-    "@progress/kendo-angular-intl": "^2.0.4",
-    "@progress/kendo-angular-l10n": "^2.0.3",
-    "@progress/kendo-angular-layout": "^5.1.0",
-    "@progress/kendo-angular-pager": "^0.2.3",
-    "@progress/kendo-angular-pdf-export": "^2.1.0",
-    "@progress/kendo-angular-popup": "^3.0.7",
-    "@progress/kendo-angular-progressbar": "^0.2.4",
-    "@progress/kendo-angular-toolbar": "^3.2.1",
-    "@progress/kendo-angular-tooltip": "^2.1.6",
-    "@progress/kendo-angular-upload": "^6.1.0",
+    "@pdftron/webviewer-html": "^4.1.2",
+    "@pdftron/webviewer-video": "^4.17.1",
+    "@progress/kendo-angular-buttons": "^7.0.2",
+    "@progress/kendo-angular-common": "^2.0.2",
+    "@progress/kendo-angular-dateinputs": "^6.0.0",
+    "@progress/kendo-angular-dialog": "^6.0.1",
+    "@progress/kendo-angular-dropdowns": "^6.0.1",
+    "@progress/kendo-angular-editor": "^3.0.1",
+    "@progress/kendo-angular-excel-export": "^4.0.3",
+    "@progress/kendo-angular-grid": "^6.0.1",
+    "@progress/kendo-angular-inputs": "^8.0.2",
+    "@progress/kendo-angular-intl": "^3.1.2",
+    "@progress/kendo-angular-l10n": "^3.0.3",
+    "@progress/kendo-angular-layout": "^6.4.1",
+    "@progress/kendo-angular-pdf-export": "^3.0.3",
+    "@progress/kendo-angular-popup": "^4.0.4",
+    "@progress/kendo-angular-toolbar": "^5.0.0",
+    "@progress/kendo-angular-tooltip": "^3.1.4",
+    "@progress/kendo-angular-upload": "^8.0.0",
     "@progress/kendo-data-query": "^1.5.5",
-    "@progress/kendo-drawing": "^1.10.0",
-    "@progress/kendo-theme-default": "^4.38.1",
+    "@progress/kendo-drawing": "^1.16.2",
+    "@progress/kendo-licensing": "^1.2.1",
+    "@progress/kendo-theme-default": "^5.1.1",
     "@progress/kendo-ui": "^2021.2.511",
     "@syncfusion/ej2-angular-buttons": "^18.4.46",
     "@syncfusion/ej2-angular-calendars": "^18.4.48",
@@ -88,32 +86,32 @@
     "npm-force-resolutions": "0.0.3",
     "rxjs": "^6.6.7",
     "rxjs-compat": "^6.6.7",
-    "tslib": "^2.2.0",
-    "zone.js": "^0.10.3"
+    "tslib": "^2.0.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1000.8",
-    "@angular/cli": "10.0.4",
-    "@angular/compiler-cli": "10.0.6",
-    "@angular/language-service": "10.0.6",
+    "@angular-devkit/build-angular": "^13.2.0",
+    "@angular/cli": "13.2.0",
+    "@angular/compiler-cli": "13.2.0",
+    "@angular/language-service": "13.2.0",
     "@babel/compat-data": "^7.14.0",
     "@babel/preset-env": "^7.14.2",
-    "@types/jasmine": "^3.7.4",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.9",
     "@types/node": "~14.0.27",
     "babel-loader": "^8.2.2",
-    "codelyzer": "^6.0.2",
+    "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.2",
-    "karma": "~5.1.1",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.12",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "^3.0.3",
-    "karma-jasmine": "~3.3.1",
-    "karma-jasmine-html-reporter": "^1.6.0",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "protractor": "~7.0.0",
     "ts-node": "~8.10.2",
     "tslint": "~6.1.2",
-    "typescript": "3.9.7"
+    "typescript": "4.5.5"
   },
   "resolutions": {
     "@babel/preset-env": "^7.8.0"
