@@ -6,7 +6,7 @@ import { FeedbackService } from '../../../../services/feedback.service';
 import { MentionItemComponent } from '../../../../shared/components/mention-item/mention-item.component';
 import '@progress/kendo-angular-editor';
 import '@progress/kendo-ui';
-
+import { HtmlUtilites } from '../../../../../../utils/html-utilities';
 declare var kendo: any;
 
 @Component({
@@ -17,13 +17,19 @@ declare var kendo: any;
 export class CommentReplyComponent implements OnInit {
 
   @Input() public feedback: IFeedback;
-  @ViewChild('myEditor', { static: true }) editor: ElementRef;
-  @ViewChild('myMention', { static: true }) mentionItem: MentionItemComponent;
+  @ViewChild('myReplyEditor', { static: false }) replyEditor: ElementRef;
+  @ViewChild('myReply2ReplyCreate', { static: false }) reply2ReplyCreate: ElementRef;
+  @ViewChild('myReply2ReplyEditor', { static: false }) reply2ReplyEditor: ElementRef;
+  @ViewChild('myReplyEditorMention', { static: false }) replyEditorMention: MentionItemComponent;
+  @ViewChild('myReply2ReplyCreateMention', { static: false }) reply2ReplyCreateMention: MentionItemComponent;
+  @ViewChild('myReply2ReplyEditorMention', { static: false }) reply2ReplyEditorMention: MentionItemComponent;
 
-  reply: boolean = false;
+  showReplyEditor: boolean = false;
+  showReply2ReplyCreate: boolean = false;
+  showReply2ReplyEditor: boolean = false;
+  reply_comment_id: number = 0;
 
   dl: string;
-  //mentions: string[] = null;
 
   constructor(private ref: ChangeDetectorRef,
     private feedbackService: FeedbackService,
@@ -36,50 +42,259 @@ export class CommentReplyComponent implements OnInit {
     })).subscribe(e => {
       this.dl = e.dl;
     });
-    this.createReplyEditor(this.feedback.commentId);
   }
 
-  showHideReplyTextarea(event, commentId) {
+  //create reply
+
+  toggleReply2ReplyCreate(event, commentId) {
     event.stopPropagation();
+    this.showReply2ReplyCreate = !this.showReply2ReplyCreate;
+    if (this.showReply2ReplyCreate) {
+      setTimeout(() => {
+        this.createReply2ReplyCreate(commentId);
+      }, 50);
 
-    this.reply = !this.reply;
+      this.showReply2ReplyEditor = false;
+      this.showReplyEditor = false;
+      this.ref.detectChanges();
 
-    this.ref.detectChanges();
-
-    setTimeout(() => {
-      if (this.editor) {
-        kendo.jQuery(this.editor.nativeElement).data("kendoEditor").focus();
-      }
-    }, 50);
+      setTimeout(() => {
+        if (this.reply2ReplyCreate) {
+          kendo.jQuery(this.reply2ReplyCreate.nativeElement).data("kendoEditor").focus();
+        }
+      }, 50);
+    }
+    else {
+      this.ref.detectChanges();
+    }
   }
 
-  createReplyEditor(commentId) {
-    kendo.jQuery(this.editor.nativeElement).kendoEditor({
+  createReply2ReplyCreate(commentId) {
+    kendo.jQuery(this.reply2ReplyCreate.nativeElement).kendoEditor({
       value: "",
-      change: function (e) {
-        //me.textChanged(this.value());
-      },
       tools: [{ name: "insertHtml" }],
       keydown: (e) => {
         if (e.key == "Enter") {
-          //e.stopPropogation();
           let me = this;
           e.preventDefault();
-          this.onChange(e, commentId);
+          this.createReply2Reply(e, commentId);
           return false;
         }
 
-        this.mentionItem.commentKeyDown(e);
+        this.reply2ReplyCreateMention.commentKeyDown(e);
       },
-      placeholder: "Write a reply here..."
     });
 
     kendo.jQuery(".k-editor-toolbar-wrap").hide();
+  }
 
-    setTimeout(() => {
-      kendo.jQuery(this.editor.nativeElement).css("height","30px")
-      this.editor.nativeElement.focus();
-    }, 200);
+
+  createReply2Reply(event: any, commentId) {
+    let editor = kendo.jQuery(`#reply2ReplyCreate-${commentId}`).data("kendoEditor");
+    let editorComment = editor.value();
+
+    console.log(this.feedback);
+
+    if (!HtmlUtilites.isEmptyHtml(editorComment)) {
+      editorComment = this.reply2ReplyCreateMention.cleanupMentionTag(editorComment);
+      var comment: IFeedback = {
+        alertId: this.feedback.alertId,
+        alertStateId: this.feedback.alertStateId,
+        alrtNotifyHdrId: this.feedback.alrtNotifyHdrId,
+        assignedEmpCode: this.feedback.assignedEmpCode,
+        boardColId: this.feedback.boardColId,
+        boardDtlId: this.feedback.boardDtlId,
+        boardHdrId: this.feedback.boardHdrId,
+        boardId: this.feedback.boardId,
+        comment: editorComment,
+        commentId: -1,
+        csCommentId: this.feedback.csCommentId,
+        csProjectId: this.feedback.csProjectId,
+        csReplyId: this.feedback.csReplyId,
+        csReviewId: this.feedback.csReviewId,
+        custodyEnd: this.feedback.custodyEnd,
+        custodyStart: this.feedback.custodyStart,
+        documentId: this.feedback.documentId,
+        emailsent: false,
+        generatedDate: null,
+        isDraft: false,
+        isSystem: false,
+        parentId: this.feedback.commentId,
+        userCode: this.feedback.userCode,
+        userCodeCp: this.feedback.userCodeCp,
+        vcCode: this.feedback.vcCode,
+        vrCode: this.feedback.vrCode,
+        flagChecked: false,
+        mentions: this.reply2ReplyCreateMention.mentions,
+        active: false,
+        isMyComment: true
+      };
+
+      this.feedbackService.createFeedback(comment);
+    }
+
+    this.showReply2ReplyCreate = false;
+    this.ref.detectChanges();
+  }
+
+
+  //edit reply
+  toggleReplyEditor(event, commentId) {
+    event.stopPropagation();
+    this.showReplyEditor = !this.showReplyEditor;
+    if (this.showReplyEditor) {
+      setTimeout(() => {
+        this.createReplyEditor(commentId);
+      }, 50);
+
+      this.showReply2ReplyEditor = false;
+      this.showReply2ReplyCreate = false;
+      this.ref.detectChanges();
+    }
+    else {
+      this.ref.detectChanges();
+    }
+  }
+
+  toggleReply2ReplyEditor(event, commentId) {
+    event.stopPropagation();
+    this.reply_comment_id = commentId;
+    this.showReply2ReplyEditor = !this.showReply2ReplyEditor;
+    if (this.showReply2ReplyEditor) {
+      setTimeout(() => {
+        this.createReply2ReplyEditor(commentId);
+      }, 50);
+
+      this.showReply2ReplyCreate = false;
+      this.showReplyEditor = false;
+      this.ref.detectChanges();
+    }
+    else {
+      this.ref.detectChanges();
+    }
+  }
+
+  createReplyEditor(commentId) {
+    kendo.jQuery(this.replyEditor.nativeElement).kendoEditor({
+      value: "",
+      tools: [{ name: "insertHtml" }],
+      keydown: (e) => {
+        if (e.key == "Enter") {
+          let me = this;
+          e.preventDefault();
+          this.updateReply(e, commentId);
+          return false;
+        }
+        this.replyEditorMention.commentKeyDown(e);
+      },
+    });
+
+    kendo.jQuery(".k-editor-toolbar-wrap").hide();
+  }
+
+  updateReply(event: any, commentId) {
+    let editor = kendo.jQuery(`#replyEditor-${commentId}`).data("kendoEditor");
+    let editorComment = editor.value();
+    editorComment = this.replyEditorMention.cleanupMentionTag(editorComment);
+
+    var comment: IFeedback = {
+      alertId: this.feedback.alertId,
+      alertStateId: this.feedback.alertStateId,
+      alrtNotifyHdrId: this.feedback.alrtNotifyHdrId,
+      assignedEmpCode: this.feedback.assignedEmpCode,
+      boardColId: this.feedback.boardColId,
+      boardDtlId: this.feedback.boardDtlId,
+      boardHdrId: this.feedback.boardHdrId,
+      boardId: this.feedback.boardId,
+      comment: editorComment,
+      commentId: this.feedback.commentId,
+      csCommentId: this.feedback.csCommentId,
+      csProjectId: this.feedback.csProjectId,
+      csReplyId: this.feedback.csReplyId,
+      csReviewId: this.feedback.csReviewId,
+      custodyEnd: this.feedback.custodyEnd,
+      custodyStart: this.feedback.custodyStart,
+      documentId: this.feedback.documentId,
+      emailsent: false,
+      generatedDate: null,
+      isDraft: false,
+      isSystem: false,
+      parentId: this.feedback.parentId,
+      userCode: this.feedback.userCode,
+      userCodeCp: this.feedback.userCodeCp,
+      vcCode: this.feedback.vcCode,
+      vrCode: this.feedback.vrCode,
+      flagChecked: false,
+      mentions: this.replyEditorMention.mentions,
+      active: false,
+      isMyComment: true
+    };
+    this.feedbackService.updateFeedback(comment);
+    this.showReplyEditor = false;
+    this.ref.detectChanges();
+  }
+
+
+  createReply2ReplyEditor(commentId) {
+    kendo.jQuery(this.reply2ReplyEditor.nativeElement).kendoEditor({
+      value: "",
+      tools: [{ name: "insertHtml" }],
+      keydown: (e) => {
+        if (e.key == "Enter") {
+          let me = this;
+          e.preventDefault();
+          this.updateReply2Reply(e, commentId);
+          return false;
+        }
+
+        this.reply2ReplyEditorMention.commentKeyDown(e);
+      },
+    });
+
+    kendo.jQuery(".k-editor-toolbar-wrap").hide();
+  }
+
+  updateReply2Reply(event: any, commentId) {
+    let editor = kendo.jQuery(`#reply2ReplyEditor-${commentId}`).data("kendoEditor");
+    let editorComment = editor.value();
+    editorComment = this.reply2ReplyEditorMention.cleanupMentionTag(editorComment);
+    var feedback = this.feedback.replies.find((v) => v.commentId == commentId);
+
+    var comment: IFeedback = {
+      alertId: feedback.alertId,
+      alertStateId: feedback.alertStateId,
+      alrtNotifyHdrId: feedback.alrtNotifyHdrId,
+      assignedEmpCode: feedback.assignedEmpCode,
+      boardColId: feedback.boardColId,
+      boardDtlId: feedback.boardDtlId,
+      boardHdrId: feedback.boardHdrId,
+      boardId: feedback.boardId,
+      comment: editorComment,
+      commentId: feedback.commentId,
+      csCommentId: feedback.csCommentId,
+      csProjectId: feedback.csProjectId,
+      csReplyId: feedback.csReplyId,
+      csReviewId: feedback.csReviewId,
+      custodyEnd: feedback.custodyEnd,
+      custodyStart: feedback.custodyStart,
+      documentId: feedback.documentId,
+      emailsent: false,
+      generatedDate: null,
+      isDraft: false,
+      isSystem: false,
+      parentId: feedback.parentId,
+      userCode: feedback.userCode,
+      userCodeCp: feedback.userCodeCp,
+      vcCode: feedback.vcCode,
+      vrCode: feedback.vrCode,
+      flagChecked: false,
+      mentions: this.reply2ReplyEditorMention.mentions,
+      active: false,
+      isMyComment: true
+    };
+    this.feedbackService.updateFeedback(comment);
+    this.showReply2ReplyEditor = false;
+    this.ref.detectChanges();
   }
 
   getEmployeeCode(feedback) {
@@ -95,49 +310,6 @@ export class CommentReplyComponent implements OnInit {
   getEmployeeInit(feedback) {
     var init = (fullname => fullname.map((n, i) => (i == 0 || i == fullname.length - 2) && n[0]).filter(n => n).join(""))(feedback.employeeFullName.split(" "));
     return init;
-  }
-
-  onChange(event: any, commentId) {
-    let editor = kendo.jQuery("textarea[editor-data='" + commentId + "']").data("kendoEditor");
-    let editorComment = editor.value();
-
-    editorComment = this.mentionItem.cleanupMentionTag(editorComment);
-    var comment: IFeedback = {
-      alertId: this.feedback.alertId,
-      alertStateId: this.feedback.alertStateId,
-      alrtNotifyHdrId: this.feedback.alrtNotifyHdrId,
-      assignedEmpCode: this.feedback.assignedEmpCode,
-      boardColId: this.feedback.boardColId,
-      boardDtlId: this.feedback.boardDtlId,
-      boardHdrId: this.feedback.boardHdrId,
-      boardId: this.feedback.boardId,
-      comment: editorComment,
-      commentId: -1,
-      csCommentId: this.feedback.csCommentId,
-      csProjectId: this.feedback.csProjectId,
-      csReplyId: this.feedback.csReplyId,
-      csReviewId: this.feedback.csReviewId,
-      custodyEnd: this.feedback.custodyEnd,
-      custodyStart: this.feedback.custodyStart,
-      documentId: this.feedback.documentId,
-      emailsent: false,
-      generatedDate: null,
-      isDraft: false,
-      isSystem: false,
-      parentId: this.feedback.commentId,
-      userCode: this.feedback.userCode,
-      userCodeCp: this.feedback.userCodeCp,
-      vcCode: this.feedback.vcCode,
-      vrCode: this.feedback.vrCode,
-      flagChecked: false,
-      mentions: this.mentionItem.mentions,
-      active: false
-    };
-
-    this.feedbackService.createFeedback(comment);
-
-    this.reply = !this.reply;
-    this.ref.detectChanges();
   }
 }
 
