@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { CommentViewComponent } from './comment-view.component';
 
@@ -6,7 +6,7 @@ describe('CommentViewComponent', () => {
   let component: CommentViewComponent;
   let fixture: ComponentFixture<CommentViewComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ CommentViewComponent ]
     })
