@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { MentionItemComponent } from './mention-item.component';
 
@@ -6,7 +6,7 @@ describe('MentionItemComponent', () => {
   let component: MentionItemComponent;
   let fixture: ComponentFixture<MentionItemComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ MentionItemComponent ]
     })
