@@ -8,6 +8,8 @@ using Microsoft.AspNetCore.Mvc.Filters;
 using Newtonsoft.Json.Linq;
 using System.Linq;
 using Webvantage.Angular.Controllers.Shared;
+using Webvantage.Angular.Models;
+using AdvantageFramework.Core.Database.Classes;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -34,13 +36,19 @@ namespace Webvantage.Angular.Controllers.Proofing
 
         // GET: api/<MarkupController>
         [HttpGet]
-        public AdvantageFramework.Core.Database.Entities.ProofingMarkup[] Get(string dl)
+        public ProofingMarkupDto[] Get(string dl)
         {
+            if (string.IsNullOrEmpty(dl))
+            {
+                Debug.WriteLine("dl is null or empty");
+                return null;
+            }
+
             try
             {
                 AdvantageFramework.Core.Web.QueryString qs = AdvantageFramework.Core.Web.QueryString.FromEncrypted(dl);
 
-                AdvantageFramework.Core.Database.Entities.ProofingMarkup[] Annotations = AdvantageFramework.Core.BLogic.Proofing.Methods.GetAnnotations(qs);
+                ProofingMarkupDto[] Annotations = AdvantageFramework.Core.BLogic.Proofing.Methods.GetAnnotations(qs);
 
                 return Annotations;
             }
@@ -53,13 +61,13 @@ namespace Webvantage.Angular.Controllers.Proofing
 
         // GET api/<MarkupController>/5
         [HttpGet("{id}")]
-        public AdvantageFramework.Core.Database.Entities.ProofingMarkup[] Get(int id, [FromQuery] string dl)
+        public ProofingMarkupDto[] Get(int id, [FromQuery] string dl)
         {
             try
             {
                 AdvantageFramework.Core.Web.QueryString qs = AdvantageFramework.Core.Web.QueryString.FromEncrypted(dl);
 
-                AdvantageFramework.Core.Database.Entities.ProofingMarkup[] Annotations = AdvantageFramework.Core.BLogic.Proofing.Methods.GetAnnotations(qs, id);
+                ProofingMarkupDto[] Annotations = AdvantageFramework.Core.BLogic.Proofing.Methods.GetAnnotations(qs, id);
 
                 return Annotations;
             }
@@ -72,30 +80,35 @@ namespace Webvantage.Angular.Controllers.Proofing
 
         // POST api/<MarkupController>
         [HttpPost]
-        public ActionResult<AdvantageFramework.Core.Database.Entities.ProofingMarkup []> Post([FromQuery] string dl, [FromBody] AdvantageFramework.Core.Database.Classes.ProofingMarkup [] Annotations)
+        public ActionResult<AdvantageFramework.Core.Database.Entities.ProofingMarkup[]> Post([FromQuery] string dl, [FromBody] ProofingMarkupDto[] Annotations)
         {
-            AdvantageFramework.Core.Database.Entities.ProofingMarkup [] newMarkups;
+            AdvantageFramework.Core.Database.Entities.ProofingMarkup[] newMarkups;
             try
             {
                 bool sendEmail = false;
+                bool isProofingMarkupComment = true;
+                bool onlyAtMentions = false;
+                bool hasAtMentions = false;
+                int documentID = 0;
 
                 AdvantageFramework.Core.Web.QueryString qs = AdvantageFramework.Core.Web.QueryString.FromEncrypted(dl);
                 List<AdvantageFramework.Core.Database.Entities.ProofingMarkup> entities = new List<AdvantageFramework.Core.Database.Entities.ProofingMarkup>();
 
-                foreach (AdvantageFramework.Core.Database.Classes.ProofingMarkup Annot in Annotations)
+                foreach (ProofingMarkupDto Annot in Annotations)
                 {
+                    //this is where the document id is set, so that it is either the one that is passed or the one from the qs
+                    documentID = Annot.DocumentId != null ? (int)Annot.DocumentId : qs.DocumentID;
                     entities.Add(new AdvantageFramework.Core.Database.Entities.ProofingMarkup()
                     {
                         //AlertId = qs.AlertID,
-                        //this is where the document id is set, so that it is either the one that is passed or the one from the qs
-                        DocumentId = Annot.DocumentId != null ? (int)Annot.DocumentId : qs.DocumentID,
-                        EmpCode = string.IsNullOrWhiteSpace(qs.EmployeeCode) ? null: qs.EmployeeCode,
+                        DocumentId = documentID,
+                        EmpCode = string.IsNullOrWhiteSpace(qs.EmployeeCode) ? null : qs.EmployeeCode,
                         Generated = DateTime.Now,
                         Markup = Annot.Markup,
                         MarkupTypeId = Annot.MarkupTypeId,
                         MarkupXml = Annot.MarkupXml,
                         Thumbnail = Annot.Thumbnail,
-                        ProofingXReviwerId = qs.ProofingStatusExternalReviewerID <= 0 ? (int ?)null : qs.ProofingStatusExternalReviewerID
+                        ProofingXReviwerId = qs.ProofingStatusExternalReviewerID <= 0 ? (int?)null : qs.ProofingStatusExternalReviewerID
                     });
                 }
 
@@ -103,11 +116,15 @@ namespace Webvantage.Angular.Controllers.Proofing
 
                 if (Annotations[0].Mentions != null && Annotations[0].Mentions.Length > 0)
                 {
-                    sendEmail = true;
+                    hasAtMentions = true;
                     _controller.AddAlertMentions(qs, qs.AlertID, Annotations[0].Mentions, newMarkups[0].CommentId != null ? (int)newMarkups[0].CommentId : 0);
+                    sendEmail = true;
+                    onlyAtMentions = true;
                 }
 
-                NotifyAlertRecipients(qs, qs.AlertID, true, true, false, false, null, true, qs.DocumentID, sendEmail);
+                NotifyAlertRecipients(qs, qs.AlertID, true, true, false, false, null, true, documentID,
+                    sendEmail, isProofingMarkupComment, onlyAtMentions, Annotations[0].Mentions, qs.EmployeeCode);
+
             }
             catch (Exception ex)
             {
@@ -121,7 +138,7 @@ namespace Webvantage.Angular.Controllers.Proofing
 
         // PUT api/<MarkupController>/5
         [HttpPut("{id}")]
-        public ActionResult<AdvantageFramework.Core.Database.Entities.ProofingMarkup> Put(int id, [FromQuery] string dl, [FromBody] AdvantageFramework.Core.Database.Classes.ProofingMarkup value)
+        public ActionResult<AdvantageFramework.Core.Database.Entities.ProofingMarkup> Put(int id, [FromQuery] string dl, [FromBody] ProofingMarkupDto value)
         {
             AdvantageFramework.Core.Database.Entities.ProofingMarkup newMarkup;
 
@@ -129,9 +146,9 @@ namespace Webvantage.Angular.Controllers.Proofing
             {
                 AdvantageFramework.Core.Web.QueryString qs = AdvantageFramework.Core.Web.QueryString.FromEncrypted(dl);
 
-                newMarkup = AdvantageFramework.Core.BLogic.Proofing.Methods.UpdateAnnotation(qs,id ,new AdvantageFramework.Core.Database.Entities.ProofingMarkup() 
+                newMarkup = AdvantageFramework.Core.BLogic.Proofing.Methods.UpdateAnnotation(qs, id, new AdvantageFramework.Core.Database.Entities.ProofingMarkup()
                 {
-                    CommentId = value.CommentID,
+                    CommentId = value.CommentId,
                     DocumentId = value.DocumentId != null ? (int)value.DocumentId : qs.DocumentID,
                     EmpCode = qs.EmployeeCode,
                     Generated = value.Generated,
